[fix-#11788]fix code's position

+++ dolphinscheduler-api/src/main/java/org/apache/dolphinscheduler/api/service/impl/ProcessDefinitionServiceImpl.java
@@ -856,10 +856,6 @@ public class ProcessDefinitionServiceImpl extends BaseServiceImpl implements Pro
             putMsg(result, Status.USER_NO_OPERATION_PERM);
             return result;
         }
-        if(ReleaseState.ONLINE == processDefinition.getReleaseState()){
-            putMsg(result, Status.PROCESS_DEFINE_STATE_ONLINE, processDefinition.getName());
-            return result;
-        }
         processDefinitionUsedInOtherTaskValid(processDefinition);
 
         // get the timing according to the process definition
@@ -876,6 +872,10 @@ public class ProcessDefinitionServiceImpl extends BaseServiceImpl implements Pro
                 putMsg(result, Status.SCHEDULE_CRON_STATE_ONLINE, scheduleObj.getId());
                 return result;
             }
+        }
+        if(ReleaseState.ONLINE == processDefinition.getReleaseState()){
+            putMsg(result, Status.PROCESS_DEFINE_STATE_ONLINE, processDefinition.getName());
+            return result;
         }
         int delete = processDefinitionMapper.deleteById(processDefinition.getId());
         if (delete == 0) {
